Scrollview + Flatlist
FINALLY GOT THE FREAKING SCROLLVIEW + FLATLIST WORKING!!! I had to use several ternaries depending on the action (whether someone rated or bookmarked a movie). Also came up with another usedState for the likes in that it updated the existing number of likes (mostly non-zero) when you liked it and it would subtract if you unliked it. Added in comments and ratings too and Profile pictures and borders for each. I kept the flatlist within the scrollview so everything was rendered  and you could scroll down to see more!